tests/loadtest/pyproject.toml
[project]
name = "greyeye-loadtest"
version = "0.1.0"
description = "GreyEye load testing suite — MVP (10-camera) and scale (100-camera) scenarios"
requires-python = ">=3.11"
dependencies = [
    "locust>=2.31",
    "httpx>=0.27",
    "websockets>=14.0",
    "nats-py>=2.9",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "PyJWT>=2.8",
    "rich>=13.0",
    "shared-contracts",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
